requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-seo-js==0.4.1
idna==3.11
pillow==12.0.0
psycopg2-binary==2.9.11
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
yt-dlp==2025.11.12
